+++ js/gen/app.socket.js
@@ -24,24 +24,24 @@ app.socket = {
 		var ip = app.config.get_sisbot_url();
 		if (ip == 'false') return; // stop if not set
 
-    //    console.log('ip===', ip)
+    //    app.log('ip===', ip)
     // var ip = app.collection.get(sisbot_id).get('data.local_ip');
 
 		if(ip.match(/^https?:\/\//i)) ip = ip.replace(/^https?:\/\//i, "");
 		if(ip.match(/:[0-9]+\/?$/i)) ip = ip.replace(/:[0-9]+\/?$/i, "");
 
 		if (this.reset_socket) {
-			console.log("Socket: Reset new Socket");
+			app.log("Socket: Reset new Socket");
 		} else if (this.server_ip == ip) {
 			app.trigger("socket:reconnect", null);
-			return console.log("Same Socket IP, skip recreate", ip);
+			return app.log("Same Socket IP, skip recreate", ip);
 		}
 
 		this.server_ip = ip;
-    console.log('Socket: Address', this.server_ip);
+    app.log('Socket: Address', this.server_ip);
 
 		if (self.socket) {
-			console.log("Socket: Close", self.socket);
+			app.log("Socket: Close", self.socket);
 
 			// turn off listeners
 			self.socket.off('connect');
@@ -57,7 +57,7 @@ app.socket = {
 			delete self.socket;
 		}
 
-    self.socket = io.connect( 'http://' + this.server_ip + ':3002'); //change to 3000 for download to work
+    self.socket = io.connect( 'http://' + this.server_ip + ':3002', {'reconnection': false}); //change to 3000 for download to work
 
 		self.socket.on('connect', function () {			self.on_connect();		});
     self.socket.on('reconnect', function() {        self.on_reconnect();    });
@@ -71,37 +71,43 @@ app.socket = {
 
     self.socket.emit('register', { id: sisbot_id });
 
+		console.log("Socket: reconnection ", this.socket.io.reconnection());
 		this.reset_socket = false;
 	},
+	set_reconnection: function(value) {
+		if (_.isBoolean(value)) this.socket.io.reconnection(value);
+		console.log("Socket: reconnection ", this.socket.io.reconnection());
+		console.log("Socket: is_connected", this.socket.connected, this.socket.disconnected);
+	},
   on_connect: function(socket) {
-    console.log('Socket: connect');
+    app.log('Socket: connect');
 		app.trigger("socket:connect", null);
   },
   on_reconnect: function() {
-    console.log('Socket: reconnect');
+    app.log('Socket: reconnect');
 		app.trigger("socket:reconnect", null);
   },
   on_reconnect_attempt: function() {
-    console.log('Socket: reconnect_attempt');
+    app.log('Socket: reconnect_attempt');
 		app.trigger("socket:reconnect_attempt", null);
   },
   on_disconnect: function() {
-    console.log('Socket: disconnect');
+    app.log('Socket: disconnect');
 		this.reset_socket = true;
 		app.trigger("socket:disconnect", null);
   },
   on_error: function(err) {
-    console.log('Socket: error', err);
+    app.log('Socket: error', err);
 		app.trigger("socket:error", err);
   },
   on_set: function(data) {
 		app.manager.intake_data(data);
   },
   on_erase: function(data) {
-    console.log('Socket: erase');
+    app.log('Socket: erase');
     //app.collection.remove(data.id);
   },
   on_test: function(data) {
-    console.log('Socket: test', data);
+    app.log('Socket: test', data);
   },
 };