+++ CHANGELOG.md
@@ -13,6 +13,7 @@
  - Drawing paths export using smooth bezier curves
  - Rescan button forces full rescan
  - Multiple table page spinners/buttons cleaned up
+ - Socket won't reconnect while page/app in background, reconnects on moving to foreground
 #2019-11-20 1.8.79
  - Now can add artist name to drawn tracks / auto populates from webcenter (waiting for update to webcenter).
  - Rho Start and finish now working correctly.